Проект разбит на модули
